Add support for ETags and the If-Match headers.
Adding support for ETags and If-Match headers. On the client side, we need to
know if the entity that we have requested from poky is the version that we
expect. If the client makes a GET request under /kv that contains the If-Match
header then value supplied will be compared with the current modified_at
timestamp. If they are equal the GET will be processed as usual. If they are
not equal the request will be rejected with a 412 status code.