Implement sproc via stored procedure
Instead of generating the query JVM-side, use a stored procedure for the mget
query.

Primary motivation for these changes is to use a IN/ANY super-select for the
query for performance reasons. There were concerns about complexity of
generating queries, so this generates a stored procedure call instead (an
arguable trade-off). A benefit of having this as a sproc is it can be used from
different clients as well as better access for load testing.

Timestamped queries aren't working quite right yet (where parameter is Timestamp
object). I am assuming this is JDBC type inference issue.