WIP adding simple benchmarking of jdbc-mget.
This is to help get started benchmarking different database configurations
and queries. It should also be helpful in preventing performance regressions.

As a first step, we can get data from different buckets and then randomize
the lookup of that data in fixed sized chunks. This should allow us to exercise
different parts of the indexes or partitions.

/cc @loganlinn